Addition of Baseline Window Input, ? Box update
The baseline stability function in roi_refine was unclear. Added a more thorough description for its ? box and updated the GUI to make it more intuitive. Also, added a "window" input to allow selection of the number of frames used to calculate the baseline value. This could cause some issues when loading saved settings, as a new variable has been added that will not be present in old settings files. I have updated the autosave script to allow for this, but previously saved settings files would likely cause an error.